Call free() on gg.img->data and gg.mask->data before setting to NULL.
In the case where malloc() is really GC_malloc(), setting the pointer
to NULL is necessary to avoid a crash.  In that case, free() will
really be GC_free(), and calling it beforehand causes no harm.  In the
case where malloc() is really libc malloc(), free() will really be
libc free(), and calling it before setting the pointer to NULL avoids
a memory leaks.

In all cases, this will ensure that the data pointer is passed to the
free() function corresponding to the malloc() that produced it.