Removed reference to obsolete variable 'hadtimes'
The function SFD_GetFont defines local variable 'hadtimes', set to true when CreationTime is read from an SFD file (else the file's mtime is used for the creation time).  However, this variable had migrated to a field of the same name in the SFD_GetFontMetaDataData structure when the SFD_GetFontMetaData function was created, and is now always false, meaning that CreationTime is *always* overwritten with the file's mtime.

This patch restores the code's intended behavior, which is to preserve the value of CreationTime if it exists in the SFD file, and only use the file's mtime if it does not.  The obsolete local variable 'hadtimes' is also removed.