Skip test for pkg-config. test for spiroentrypoints.h and libspiro.
Some distros are not completely configured correctly for pkg-config.
There is a school of thought that argues for using PKG_CHECK_MODULES,
but that is not necessary. The standard way to check if a library
exists and to link against it is to use AC_CHECK_LIB(), and also check
for the developer header file(s) using AC_CHECK_HEADERS().

Fixed fontforge_create_pkg-config.m4 to only include libspiro as a
dependecy ONLY if it exists AND user didn't say no to include libspiro.

This comes to a hard-stop if user wants libspiro, but it's not available.