Skip test for pkg-config. test for uninameslist.h and libuninameslist.
Some distros are not completely configured correctly for pkg-config.
There is a school of thought that argues for using PKG_CHECK_MODULES,
but that is not necessary. The standard way to check if a library
exists and to link against it is to use AC_CHECK_LIB(), and also check
for the developer header file(s) using AC_CHECK_HEADERS().

Fixed fontforge_create_pkg-config.m4 to only include libuninameslist as
a dependecy ONLY if it exists AND user didn't say no to include it.

This comes to a hard-stop if user wants libuninameslist, but it's not
available. This also tests for version 0.3 and 0.4 of libuninameslist

Libunicodenames was placed in '#' (commented) since it was going to be a
fair bit of re-write to work-around the repeated AC_DEFINE in configure.ac
Issue #1341 recommends sticking with 1 library rather than support both,
and libunicodenames seems to be limited in function access and OS choices.
Users that want to use libunicodenames can move the comment '#' to switch
libraries for one or the other (see configure.ac) since the active code
is still in place if you want to use the other library, or if there is
another library in future, shouldn't be hard to modify for others:
FONTFORGE_ARG_WITH_LIBUNINAMESLIST
#FONTFORGE_ARG_WITH_LIBUNICODENAMES