Work through the compiler warnings in ufo.c
These functions could not be declared with const SplineChar* because one
or more functions that they subsequently call modify that parameter:
* UFOOutputFontInfo
* UFOOutputFeatures
* WriteUFOFont

The most notable example is a call which leads to SFReadChar being called.
There may be other functions which also modify the parameter, but tracing
the calls is tedious.