Fix bugs in GIOGetMimeType()
Converted sniff_length into a define because this is a compiled binary.
g_content_type_guess() was wasted compute time at the front, and the
result was thrown away, so moved it to the back as a last resort if
sniffing data for mime type failed.
Accept the possibility of "content_type==NULL" after all these check.