Reveal partially hidden button in "FontInfo / Layers"
On both Ubuntu and Cygwin, when the "Layers" pane of the "Font Info..."
dialog was displayied, the "Delete" button below the "Layers:" list box
was obscured or hidden.

Upon first seeing this, the thought was that the entire "Font Info..."
dialog window would have to be made taller to display the button.
(Another pane "Lookups" would also seem to need this)

Instead, this fix merely shortens the list box, allowing the button to
rise and be seen.  This still allows as many as 8 layer descriptions
to be displayed without scrolling.