issue1344: work on source tree clones from bootstrap.
Some build environments clone the source tree before starting the
bootstrap/configure process. For example, homebrew on osx. That being
the case, some of these do not copy the .git tree in the clone
probably because it is not intended to be long lived. BUT if there is
no .git subdirectory then the old bootstrap would subtly fail by
forcing the GIT command to be out of bounds.

This update allows bootstrap to run in such .git'less clones and thus
removes one hurdle to FontForge building from homebrew.