Fix small leak in pythonui.c python_ui_setup_callback()
Another use of the recent routine gutils/fsys.c :: getFontForgeUserDir()
that requires a free() of the newly allocated string after use.

Note that there was an unprotected path if a NULL was returned from
getFontForgeUserDir().  The old code would have tried to use the null
pointer.  The fix protects against that, but rather does 'nothing' if
NULL is returned.  Maybe it should blow up horrible bad instead?