Fix leak of strings in openfontdlg.c FVOpenFont()
A list of glyph list names was created for display in the "Open Font"
dialog.  The name strings were stuffed into the dialog creation data
structures.  After the dialog was dismissed, nobody was free()ing those
name strings.

Moved the names list array free() to end of dialog routine, and first
used that list to locate and free() all the name strings.