Leaks when mouse moving over "Open Font" file list
The "Open Font" dialog was tracking mouse movements over the list of
possible filenames.  The desire was to always have the font name(s)
of files available for the balloon popup over the filename, if the
mouse paused long enough over the same filename.

For each mouse movement the filename under the cursor would be
identified and that file path generated in a newly allocated string.
That string was not free()d after use.

Once the file path was identified the font file would be opened to
discover the font name(s) in that file.  These were returned in a
list of strings, with list and strings newly allocated.  These were
not being free()d.

Depending on how much mouse movement occurred over the file list (and
apparently including movement over the file list scrollbar?) 10's to
100's of KB were leaked.

Both the found file path names and the font names are free()d by this
fix.

(Wow, if I'd'a known all these leaks were in this one area I'd'a
  created one big PR ... ?!?)