issue527: width in metrics view can now be changed many times.
No CPU lock or crashes. Width can now also be changed by dragging the
chars in the metricsview. Maybe my config is different to what it was
before but drags now work here on F20.

Also tested manually changing the width, lbearing, rbearing using the
up/down arrow keys in the grid at the bottom of the metricsview
window.