Display errors using LogError in "Warnings" popup window
More investigation found that `LogError()` seemed to fit the need to
present errors to users in a more visible way, rather than using
`fprintf(stderr, ...)`.  `LogError()` will popup a "Warnings" window
if not yet displayed when running with a UI.  When not running under
a UI, it will print to stderr instead.

Also, use of asprintf() under GNU platforms generates a complie
warning, which can be avoided by defining `_GNU_SOURCE`, but which
must be defined before several other includes, and must be defined
as '1' to agree with `Python.h`