Fix probs in BdfCRefFixup() on glyph component ref error
I found two problems with the code in BdfCRefFixup() that detects references
to missing componentxs glyphs.  The error handling code was trying to elide
the failing reference node, but would do so only if this was the first such
error detected.  Then that same code would free the node but with following
code still making references to the free()d node.

Relocated some normal code up into the non-failure clause.  Separated the error
logging code (once per parsed bitmap) from the cleanup code (on every error).

Tested somewhat in the debugger.  And yes, this *was* in the Coverity reports!