Nit - replace computing constant with that constant
parsettf.c :: readdate() was computing the time difference between
the 1904 epoch time ala Mac and the 1970 epoch time ala Unix.
However this value is a constant and need not be recomputed each time.
One could inspect the computed value once and make static, or derive
the value from various references on the web.

This change replaces code and loop with initialization of a variable
with the constant value.

Tested against three different ttf/ttc files and verified from the
Element / Font Info... / Dates displays before and after change.