Fix crash #1671 gdraw/gresedit.c
GResEditDlg() has sections to set up the different kinds of controls as
found in various panes in the X resource dialog.  In two panes clicking
on the 'inherit' checkbox for the "Font:" item would crash FFC ("Text"
and "Button" panes).

GRE_InheritFontChange() runs to react to those user clicks.  It calls
GRE_FigureInheritance() to figure out the inheritance issues (whether
to inherit from the parent GGadget type).

A parameter to GRE_FigureInheritance() 'is_font' is used to flag that a
more complicated font data item is being manipulated. That flag was not
set on this call, even though a font item *was* being manipulated, which
then caused a crash in a deeper call to inherit_font_change().

Tested and cures crash.  Inheritance appears to work correctly...

During investigation, noted that descriptive title in the "Progress"
pane was wrong.  Changed this to "Progress Bars".