Python 3 compatibility: glyph.getPosSub() glyph.removePosSub()
In FontForge built with Python 3, script APIs glyph.getPosSub()
and glyph.removePosSub() would reject all calls displaying error
messages such as:
      TypeError: expected bytes, str found

This problem not seen running FontForge built with Python 2.

The code had never been converted to be compatible with Python 3, and
was still using PyBytes_* routines to handle strings.  Changed to use
PyArg_ParseTuple() to handle the single argument.

Tested with variations on argument types and counts.  API _function_
was not tested.