Python API fails: font.size_feature (getter)
When full 'size' information is present for a font (the 5 value tuple), then
    data = font.size_feature
would fail with an internally-generated Python exception complaining:
      IndexError: tuple assignment index out of range

The code was calling PyTuple_SetItem() using the wrong variable to index into
the new tuple.

This API was unlikely to have ever worked when full 'size' information
was present.  When language info was absent, then only a single value
'design_size' was correctly returned using a different code path.

However, this GPOS 'size' feature apparently had little industry acceptance,
and so not working was likely not a problem...?

Tested with several different 'size' values as set during testing of the setter.