Python 3 compatibilty error: glyph.glyphclass setter
In FontForge built with Python 3, script API glyph.glyphclass=...
would reject calls to set a new glyph class name string, with error:
      TypeError: expected bytes, str found

This problem not seen running FontForge built with Python 2.

The code had not been converted to be compatible with Python 3 strings
and was still using the Python 2-compatible PyBytes_* APIs.

Tested with variations on argument types and values.