Coverity parsepdf.c toknames[] PDF parsing bug!
Coverity for the win!

Coverity pointed out a typo affecting the contents of an array used when
parsing PDF drawing commands. Astonishingly, the interpretation of some
less-used PDF commands has been wrong since 'forever'.

If there have ever been problems with processing 'some' PDFs, including
PDFs generated by FontForge (!), this bug could have contributed.

An array toknames[] is used to lookup drawing command names.  Indexes
within this table are correlated with the enum constants pstoks.

The typo meant that two name strings, which were intended to be the
separate commands
>   B   Fill and stroke the path (non-zero winding rule)
>   B*  Fill and stroke the path (even-odd winding rule)
were accidentally concatenated as "BB*", and so occupied only one slot
in the table.  This not only meant these two commands were not recognized,
but that every command after that in the table was assigned the *wrong*
enum value.  Such that we had wonderful nonsense like
>       "false"  ==  pt_true

It appears that the table was sorted so that the most frequently seen
commands occurred earlier in the table.  This is the only way that 'enough'
parsing worked so that the bug was not forced to the surface.

Reported as
> CID 1225991 (#1 of 1): Missing comma in a string array initialization (MISSING_COMMA)
> missing_comma: In the initialization of toknames, a suspicious concatenated string "BB*" is produced.

Tested with FontForge-generated PDFs, which happen to contain a command
'D1', which was one of those affected.  Previously matched as index 23 ==
pt_paintnoop it is now correctly matched as index 24 == pt_setcachedevice

Also the dumped table toknames[] previously showed
>       "f*",   "BB*",  "b",
and now shows
>       "f*",   "B",    "B*",
>       "b",    "b*",   "n",