Warn for PDF's with compression + unsup'd filters
As described in issue #1721, some PDF's will crash FontForge due to the
use of added filters in combination with object stream compression.

In the example crash, FlateDecode compression was combined with PNG value
filtering,
>   /Filter /FlateDecode
>   <</DecodeParms<</Columns 5/Predictor 12>>

FontForge did not look for nor handle the /DecodeParms value for any
objects.  This basically meant that the object stream was corrupted
and unguarded code used the resulting bad values causing the crash.

Filter implementation would require adding a step after decompression
to un-filter the stream.  There are a number of known filters.  (see
PDF version 1.7 document section 3.3 "Filters" for more details)

As implementation would take time and sample PDF files not available
at present, for now just warn the user that this PDF file contains
features not supported and don't use the file (preventing crash).

The warning as seen from the UI, in "Warnings" popup:
> Unsupported decode filter parameters : <</Columns 5/Predictor 12>>

There may be other warning/error messages, depending on what data part
of the PDF file is found to have this unsupported combination.

Note that this change's check for the problem situation is rather
superficial.  FontForge can be forced to handle subdictionary values
but it is very painful.  Better would be to completely unpack the
/DecodeParms subdictionary and verify that the Predictor value is
not '1'. As it is we only check if /DecodeParms is present and then
give up and warn.