parsepdf.c FindXref() small bug finding startxref
While examining code in parsepdf.c I was confused by code in routine
FindXref() that locates the section in a PDF file that points to the
start of the tables in a PDF. Checking more closely and comparing with
the previous version of the routine, I realized it could be tripped up
by slightly malformed trailers, e.g.
>     endobj
>     sstartxref
>     2250179
>     %%EOF

At first I reworked the code, fusing new code with old to correct the
glitch. But then I realized the code using loops was confusing without
needing to be in order to work.  Replaced loops with straightforward
reading of needed algorithm.

Tested with both good and multiple constructed 'bad' PDF file trailers.

This problem has not actually been demonstrated to my knowledge.
Hitting the bug would have produced messages like:
> Doesn't look like a valid pdf file, couldn't find xref section
> Couldn't find a font file named /home/tom/study/fonts/FF/gdb.pdf.test.3.pdf
> gdb.pdf.test.3.pdf is not in a known format (or uses features of that format fontforge does not support, or is so badly corrupted as to be unreadable)