Remove HelpDir preference
There’s no need for this: help should always be installed in the right
place for the given build. We should not encourage FF to be installed
without help files, even if this will sometimes happen (e.g. when it’s
packaged, but in that case I’d expect a corresponding -doc package that
installs the help in the right place).

(Suppose a user on an increasingly-mythical multiuser system whose
sysadmin has installed FF but not the help files: in that case, let that
user install their own copy of FF, then they can be sure the help files
they install will actually match it.)